backend/requirements.txt
praw>=7.7.1
PyYAML>=6.0.2
pydantic>=2.7.0
pytest>=7.4.0
openai>=1.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
apscheduler>=3.10.4
